moving things to `src/suql`
