+++ src/suql/sql_free_text_support/prompts/field_classification.prompt
